% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxw_full_query_as_list}
\alias{pxw_full_query_as_list}
\title{Return full query as a list}
\usage{
pxw_full_query_as_list(url, time_all = TRUE)
}
\arguments{
\item{url}{A URL to the table (web interface or API URL).}

\item{time_all}{If TRUE (default), the time variable is set to \code{c("*")}.}
}
\value{
A list containing the full query.
}
\description{
This function evaluates and returns the full query structure as a list.
}
