% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxw_prepare_full_query}
\alias{pxw_prepare_full_query}
\title{Prepare a query for printing}
\usage{
pxw_prepare_full_query(url, time_all = TRUE)
}
\arguments{
\item{url}{A URL to the table (web interface or API URL).}

\item{time_all}{If TRUE (default), the time variable is set to \code{c("*")}.}
}
\value{
A list containing the full PxWeb query structure.
}
\description{
This function prepares the full query structure for a given PxWeb URL.
}
