% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxw_print_full_query}
\alias{pxw_print_full_query}
\title{Print full query from a table URL}
\usage{
pxw_print_full_query(url, time_all = TRUE)
}
\arguments{
\item{url}{A URL to the table (web interface or API URL).}

\item{time_all}{If TRUE (default), the time variable is set to \code{c("*")}.}
}
\description{
This function prints the full PxWeb query structure for a given URL.
}
\examples{
pxw_print_full_query(url = "https://pxdata.stat.fi/PxWeb/api/v1/fi/StatFin/ati/statfin_ati_pxt_11zt.px")
pxw_print_full_query(url = "https://pxdata.stat.fi/PxWeb/api/v1/fi/StatFin/tyokay/statfin_tyokay_pxt_115u.px")
}
